Move step logic to questionnaire reducer
For #1691
