+++ web/static/js/components/ui/helpers.jsx
@@ -17,16 +17,6 @@ export const roleDisplayName = (role) => {
   }
 }
 
-export const canBeRelevant = stepType => {
-  switch (stepType) {
-    case 'multiple-choice':
-    case 'numeric':
-      return true
-    default:
-      return false
-  }
-}
-
 export const dayLabel = (day) => {
   switch (day) {
     case 'sun': return i18n.t('Sun')
